+++ Creature.cpp
@@ -1,0 +1,10 @@
+#include "Creature.h"
+int Creature::pos() const
+{
+	return pos_;
+}
+
+void Creature::setPos(int position)
+{
+	pos_ = position;
+}
