+++ Player.h
@@ -1,0 +1,6 @@
+#ifndef Player_h
+#define Player_h
+#include "Creature.h"
+class Player : public Creature {
+};
+#endif
